overlay 65: decompile the trade screen's summary of a Pokemon, matching
ov65_0221D674 draws one side's chosen Pokemon on the wireless trade
screen: its picture, a second sprite set from its record, its nickname,
its gender mark (none for an egg or an unnamed Nidoran), its level (none
for an egg) and the name of its held item, read into a String_New(20). That string is where the open "item names longer than
their buffers" row is left: sixteen of hg-engine's item names are past 19
characters. The function comes out of the assembly first, on its own.

overlay_65.s is split after it: the rest of the overlay goes to
asm/overlay_65_0221D8C4.s, and src/trade_select_mon_panel.c is linked
between the two. The C describes only what the function reads of the
overlay's state (the message bank at 0x190, a string at 0x19C, the
sprites at 0x40C and the twelve 16-byte Pokemon records at 0x69C) and of
its picture table ov65_0221FEA4, now public in the assembly. Sprite_SetFlipMode
takes an int: the caller passes a u16 with no truncation, which a u8
parameter would have forced; its two other callers and the function
itself compile to the same bytes.

Engine layer (retail code). HeartGold and SoulSilver build byte for byte
what they did before (4242d1fa, cfb3e679), and the link map puts the
function in the new object.

+++ include/sprite.h
@@ -125,7 +125,7 @@ void Sprite_SetDrawFlag(Sprite *sprite, BOOL flag);
 void Sprite_SetAnimActiveFlag(Sprite *sprite, BOOL flag);
 void Sprite_SetAnimSpeed(Sprite *sprite, fx32 frame);
 void Sprite_SetAffineOverwriteMode(Sprite *sprite, u8 affine);
-void Sprite_SetFlipMode(Sprite *sprite, u8 flip);
+void Sprite_SetFlipMode(Sprite *sprite, int flip);
 VecFx32 *Sprite_GetMatrixPtr(Sprite *sprite);
 VecFx32 *Sprite_GetScalePtr(Sprite *sprite);
 u16 Sprite_GetRotation(Sprite *sprite);
